codeflix/catalog-admin/infrastructure/src/main/java/com/tresende/catalog/admin/infrastructure/genre/GenreMySQLGateway.java
package com.tresende.catalog.admin.infrastructure.genre;

import com.tresende.catalog.admin.domain.genre.Genre;
import com.tresende.catalog.admin.domain.genre.GenreGateway;
import com.tresende.catalog.admin.domain.genre.GenreID;
import com.tresende.catalog.admin.domain.pagination.Pagination;
import com.tresende.catalog.admin.domain.pagination.SearchQuery;
import com.tresende.catalog.admin.infrastructure.genre.persistence.GenreJpaEntity;
import com.tresende.catalog.admin.infrastructure.genre.persistence.GenreRepository;
import org.springframework.stereotype.Component;

import java.util.List;
import java.util.Objects;
import java.util.Optional;

@Component
public class GenreMySQLGateway implements GenreGateway {

    private final GenreRepository genreRepository;

    public GenreMySQLGateway(final GenreRepository genreRepository) {
        this.genreRepository = Objects.requireNonNull(genreRepository);
    }

    private Genre save(final Genre aGenre) {
        return this.genreRepository.save(GenreJpaEntity.from(aGenre)).toAggregate();
    }

    @Override
    public Genre create(final Genre aGenre) {
        return save(aGenre);
    }

    @Override
    public void deleteById(final GenreID anId) {
        final var genreId = anId.getValue();
        if (genreRepository.existsById(genreId)) {
            genreRepository.deleteById(genreId);
        }
    }

    @Override
    public Optional<Genre> findById(final GenreID anId) {
        return Optional.empty();
    }

    @Override
    public Genre update(final Genre aGenre) {
        return save(aGenre);
    }

    @Override
    public Pagination<Genre> findAll(final SearchQuery aQuery) {
        return null;
    }

    @Override
    public List<GenreID> existsByIds(final Iterable<GenreID> ids) {
        return List.of();
    }
}
